# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "School Backend API Documentation",
        "version": "1.0"
    },
    "paths": {
...
        "/api/user/profile/facilities": {
            "get": {
                "tags": [
                    "Facilities"
                ],
                "summary": "Get all facilities",
                "description": "Retrieve all facilities",
                "operationId": "getAllFacilities",
                "responses": {
                    "200": {
                        "description": "Data ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data ditemukan"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/FasilitasResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user/profile/facilities/{id}": {
            "get": {
                "tags": [
                    "Facilities"
                ],
                "summary": "Get specific facility",
                "description": "Retrieve a specific facility by its ID",
                "operationId": "getFacilityById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data ditemukan"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/FasilitasResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "FasilitasResource": {
                "properties": {
                    "id_facility": {
                        "type": "integer",
                        "example": 1
                    },
                    "facility_name": {
                        "type": "string",
                        "example": "Lab Komputer"
                    },
                    "facility_image": {
                        "type": "string",
                        "example": "img/fasilitas/lab_komputer.png"
                    },
                    "prodi": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "nama_prodi": {
                                "type": "string",
                                "example": "Teknik Informatika"
                            },
                            "prodi_short": {
                                "type": "string",
                                "example": "TI"
                            }
                        },
                        "type": "object"
                    },
                    "facility_text": {
                        "type": "string",
                        "example": "Laboratorium komputer dengan fasilitas lengkap."
                    }
                },
                "type": "object"
            },
...
        }
    }
}